+++ packages/ListBox/CHANGELOG.md
@@ -10,4 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add onKeyUp event for `SPACE`, `ESC` and `ENTER` before onKeyDown was handling all key interaction now its only responsible for `↑`,`↓` keys.
   this might break testing that depends onKeyDown event.
-- Add `customLabel` prop on `ListBox.Trigger` component to support custom text as the label.
+
+## [0.9.3] - 2020-08-13
+
+### Changed
+
+- Add `label` prop on `ListBox.Trigger` component to support custom text as the label.